requirements.txt
# Web Framework
Flask==2.3.3
Werkzeug==2.3.7
flask-cors==4.0.0
python-dotenv==1.0.0

# Image Processing
Pillow==10.0.0
opencv-python==4.8.0.76
numpy==1.25.2
scipy==1.11.2

# 3D Processing
open3d==0.17.0
trimesh==3.23.5
PyMeshLab==2022.2.post3

# Deep Learning & AI
torch==2.0.1
torchvision==0.15.2
transformers==4.33.1
sentence-transformers==2.2.2

# Chatbot & NLP
nltk==3.8.1
spacy==3.6.1
python-dateutil==2.8.2

# Utilities
requests==2.31.0
tqdm==4.66.1
pandas==2.1.0
matplotlib==3.7.2

